core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@5116 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/persist/TableSpaceManager.java
@@ -50,7 +50,7 @@ public interface TableSpaceManager {
     /**
      * Allocates a block
      */
-    long getFilePosition(int rowSize, boolean asBlocks);
+    long getFilePosition(long rowSize, boolean asBlocks);
 
     boolean hasFileRoom(int blockSize);
